agents/google-adk/requirements.txt
google-adk>=1.31.0
google-genai>=1.0.0
python-dotenv>=1.0.0
